requirements.txt
python-dotenv
pandas
numpy
pymongo
certifi
pymongo[srv]
dill
pyaml

##-e . 
